+++ sdk/async_api/anonymous_schema106.py
@@ -1,4 +1,4 @@
 from enum import Enum
 
-class AnonymousSchema95(Enum): 
+class AnonymousSchema106(Enum): 
   MARKET_TYPE_MISMATCH = "MarketTypeMismatch"