+++ sdk/async_api/anonymous_schema121.py
@@ -1,4 +1,4 @@
 from enum import Enum
 
-class AnonymousSchema110(Enum): 
+class AnonymousSchema121(Enum): 
   FILL_EXCEEDS_ORDER_BASE_DELTA = "FillExceedsOrderBaseDelta"