+++ sdk/async_api/anonymous_schema148.py
@@ -1,4 +1,4 @@
 from enum import Enum
 
-class AnonymousSchema137(Enum): 
+class AnonymousSchema148(Enum): 
   SAME_ACCOUNT_ID = "SameAccountId"