+++ sdk/async_api/anonymous_schema150.py
@@ -1,4 +1,4 @@
 from enum import Enum
 
-class AnonymousSchema139(Enum): 
+class AnonymousSchema150(Enum): 
   DECODED_REASON = "DecodedReason"