+++ sdk/async_api/anonymous_schema69.py
@@ -1,4 +1,4 @@
 from enum import Enum
 
-class AnonymousSchema58(Enum): 
+class AnonymousSchema69(Enum): 
   UNAUTHORIZED = "Unauthorized"