+++ sdk/async_api/anonymous_schema71.py
@@ -1,4 +1,4 @@
 from enum import Enum
 
-class AnonymousSchema60(Enum): 
+class AnonymousSchema71(Enum): 
   UNAUTHORIZED_SIGNER = "UnauthorizedSigner"