+++ sdk/async_api/anonymous_schema88.py
@@ -1,4 +1,4 @@
 from enum import Enum
 
-class AnonymousSchema77(Enum): 
+class AnonymousSchema88(Enum): 
   PRICE_DEVIATION_TOO_LARGE = "PriceDeviationTooLarge"