+++ sdk/async_api/anonymous_schema94.py
@@ -1,0 +1,4 @@
+from enum import Enum
+
+class AnonymousSchema94(Enum): 
+  ACCOUNT_BELOW_IM = "AccountBelowIM"